Change 'Board editors name' & fix active boards display in user profile
This commit adds new migration file which changed 'Board editors' group
to 'This app editors' so the name of the group matches the board better that
it was supposed to be from the start.

Additionally fixed display of "Active in" on user profile where user boards
were printed based on the number of user boards. This was achieved by modyfing
the user_profile view with more robust solution by passing user boards to context
rather than reading the groups directly from request.user in template.